fix syntax errors in Makefile
Calls to $(error) or $(warning) that are indented with tabs generate
gmake "recipe commences before first target" errors.   example: the
warning call on line 121 of buildflags.mak does this:

% make PSM_DISABLE_AVX2=yes
/tmp/bob/opa-psm2/buildflags.mak:121: *** recipe commences before first target.  Stop.
%

To resolve you can delete the spaces, turn the tabs to spaces, or
prefix error/warning calls with an assignment ("anerr :=").  Since
the "anerr" was already in use in other places in psm2, apply that fix
to the rest of the Makefile and buildflags.mak.

Also, the definition of DISTRO assumes /bin/sh is bash and uses
the bash "[[ a == b ]]" syntax.  This generates a shell error
every time you run make if /bin/sh isn't bash.  For example,
on an ubuntu box:

% make
/bin/sh: 1: [[: not found
...

Resolve this by converting "[[ a == b ]]" to the more portable
"[ a = b ]" format.   This resolves opa-psm2 issue #55 ...